requirements.txt
numpy
Flask
tensorflow
gunicorn
python-twitter
six
